Remove previous members from team page
We are super thankful for all the people who have built Portmaster with us, but it is getting clearer that people are getting the wrong idea from us showing previous members on the website. This is why we decided to remove them. Last week, the legal part of the departure of our Co-Founder David was completed, and we wanted to wait until this was done because technically he was still part of Safing until then, just not in the operational daily business.